Add features and fix bugs
1.修复碰撞检测路径不准与播放碰撞特效位置的问题；
2.手机版增加技能拖拽施法，与提示功能；
3.手机跳跃增加点击朝着面前跳跃；
4.多线程加载，增加线程安全相关；
5.使用智能指针；
6.优化资源内存占用；
7.其它功能与bug修复。

+++ src/File/PakFile.cpp
@@ -87,7 +87,7 @@ int PakFile::findFileInPak(unsigned int fileID, const std::string & pakName)
 	if (fp)
 	{
 		SDL_RWseek(fp, 0, 2);
-		int size = SDL_RWtell(fp);
+		int size = (int)SDL_RWtell(fp);
 		SDL_RWseek(fp, 0, 0);
 		if (size >= sizeof(PakHead))
 		{
@@ -315,7 +315,7 @@ bool PakFile::readPak(const std::string & pakName, int index, std::unique_ptr<ch
 		if (fp)
 		{
 			SDL_RWseek(fp, 0, 2);
-			int size = SDL_RWtell(fp);
+			int size = (int)SDL_RWtell(fp);
 			SDL_RWseek(fp, 0, 0);
 			PakHead pakHead;
 			SDL_RWread(fp, &pakHead, sizeof(PakHead), 1);
